chore: full repo cleanup + bug fixes + wire AI generator
Bug fixes:
- Fix CRITICAL: COALESCE(p.acreage, p.lot_acres) referenced non-existent
  column, crashing GET /api/properties and GET /api/properties/:id — changed
  to p.acreage AS lot_acres in both queries
- Remove dead /admin-assets static route pointing to missing admin/ directory
- Add slugify to import in routes/api.js (was using inline require as workaround)
- Fix requireApiKey path guard from '/api/health' to '/health'
- Fix null price showing '$0' on listing cards (now shows 'Contact for price')
- Fix escapeHtml in listing.js to handle null values (String(s ?? ''))

Cleanup:
- Remove 13 stale git worktrees and .claire/ directory
- Remove empty scaffolding dirs: config/, api/services/, data/, reports/,
  uploads/, tools/
- Remove root package.json and package-lock.json (Railway uses railway.json)
- Remove root .env.example (api/.env.example is authoritative)
- Delete stale/contradictory docs: PROJECT.md, AGENTS.md, AI_LISTING_MACHINE.md
- Replace PostgreSQL database/schema.sql with accurate SQLite reference
- Rewrite CONTEXT.md to describe actual stack (SQLite, route locations,
  column names, rules for AI assistants)
- Rewrite README.md with accurate structure, full route tables, deploy notes
- Rewrite api/.env.example as single annotated source of truth
- Fix .gitignore: listings/ rule now tracks _template, excludes .db.bak files
- Fix api-docs-consistency.yml: was grepping server.js for routes that moved
  to routes/*.js; now greps the correct files
- Update .github/instructions and copilot-instructions to reflect routes/*.js
  structure and acreage/lot_acres column rule

Features:
- Wire ai-generator.js into admin panel: GET /admin/ai/:id shows full
  generated content with per-section copy buttons; POST /admin/ai/:id
  triggers GPT-4o generation and saves to DB + listing.json
- Add purple 🤖 AI button to each row in the listings dashboard
