--- .claire/worktrees/gifted-lalande/app/styles.css
@@ -1,1 +1,0 @@
-/* intentional typo test */
