--- .env.example
@@ -1,48 +1,0 @@
-# ── Server ────────────────────────────────────────────────
-PORT=3001
-NODE_ENV=development
-# CRITICAL on Railway/Render: point this to a mounted persistent volume path
-# Railway: create a Volume → mount at /data → set DATABASE_PATH=/data/wv_property.db
-DATABASE_PATH=./database/wv_property.db
-SESSION_SECRET=change-me-session-secret
-ADMIN_PASSWORD=change-me-admin-password
-
-# ── API Auth ──────────────────────────────────────────────
-API_KEY=your-secret-api-key
-
-# ── Google OAuth2 (Gmail + Drive — existing flow) ─────────
-GOOGLE_CLIENT_ID=
-GOOGLE_CLIENT_SECRET=
-GOOGLE_REFRESH_TOKEN=
-
-# ── Gmail notifications ───────────────────────────────────
-GOOGLE_GMAIL_USER=malickland@gmail.com
-NOTIFICATION_EMAIL=phil@malickland.net
-
-# ── Google Service Account (Sheets + Drive service) ───────
-# Option 1: inline JSON string
-GOOGLE_SERVICE_ACCOUNT_KEY={"type":"service_account","project_id":"your-project","private_key_id":"...","private_key":"-----BEGIN RSA PRIVATE KEY-----\n...\n-----END RSA PRIVATE KEY-----\n","client_email":"your-sa@project.iam.gserviceaccount.com","client_id":"...","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token"}
-# Option 2: path to service account JSON file
-# GOOGLE_APPLICATION_CREDENTIALS=./secrets/service-account.json
-
-# ── Google Drive ──────────────────────────────────────────
-GOOGLE_DRIVE_FOLDER_ID=
-
-# ── Google Sheets ─────────────────────────────────────────
-LISTINGS_SHEET_ID=your-listings-spreadsheet-id
-LEADS_SHEET_ID=your-leads-spreadsheet-id
-
-# ── OpenAI (listing description generation) ──────────────
-OPENAI_API_KEY=
-
-# ── Anthropic / Claude (chatbot widget) ───────────────────
-ANTHROPIC_API_KEY=
-
-# ── Analytics ─────────────────────────────────────────────
-GA_MEASUREMENT_ID=G-XXXXXXXXXX
-
-# ── Slack (lead notifications) ────────────────────────────
-SLACK_WEBHOOK_URL=
-
-# ── Site ──────────────────────────────────────────────────
-SITE_URL=https://malickland.net
