+++ .github/instructions/ci.instructions.md
@@ -9,4 +9,4 @@ applyTo: ".github/workflows/**"
 - API docs consistency check (.github/workflows/api-docs-consistency.yml) is required — do not remove or disable it.
 - Do not add workflows that bypass branch protection or skip required status checks.
 - Secrets must be referenced via ${{ secrets.SECRET_NAME }} — never hardcoded.
-- New CI jobs that parse api/server.js should use grep patterns consistent with the existing api-docs-consistency.yml workflow.
+- New CI jobs that parse routes should grep api/routes/*.js — routes are no longer defined inline in api/server.js.
