+++ .github/workflows/codescan.yml
@@ -29,14 +29,14 @@ jobs:
         runs-on: ubuntu-latest
         steps:
             -   name: Checkout repository
-                uses: actions/checkout@v4
+                uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4
             -   name: Set up Java 17
-                uses: actions/setup-java@v4
+                uses: actions/setup-java@c1e323688fd81a25caa38c78aa6df2d33d3e20d9 # v4
                 with:
                     distribution: temurin
                     java-version: '17'
             -   name: Cache files
-                uses: actions/cache@v4
+                uses: actions/cache@0057852bfaa89a56745cba8c7296529d2fc39830 # v4
                 with:
                     path: |
                         ~/.sonar
@@ -50,6 +50,6 @@ jobs:
                     projectKey: ${{ secrets.CODESCAN_PROJECT_KEY }}
                     generateSarif: "1"
             -   name: Upload SARIF file
-                uses: github/codeql-action/upload-sarif@v3
+                uses: github/codeql-action/upload-sarif@865f5f5c36632f18690a3d569fa0a764f2da0c3e # v3
                 with:
                     sarif_file: codescan.sarif
