--- PROJECT.md
@@ -1,24 +1,0 @@
-# MalickLand / WV Property Intelligence — PROJECT STATE
-
-## OBJECTIVE
-Build a fast, land-focused real estate platform with:
-- listings system
-- admin panel
-- public website
-- Google Sheets backend
-
-## CURRENT STATUS
-- Server running (Express 5 fixed)
-- Admin route configured
-- API routes missing
-
-## CURRENT TASK
-Fix and verify:
-- /admin loads
-- correct file path is used
-
-## NEXT TASK
-Build listings pipeline:
-- schema
-- API routes
-- Google Sheets integration
