+++ api/.env.example
@@ -1,26 +1,41 @@
-# ── Server ───────────────────────────────────────────────────
-PORT=3000
-ADMIN_PASSWORD=wvrea2026
+# ── Server ─────────────────────────────────────────────────────────────────────
+PORT=3001
+NODE_ENV=development
+
+# ── Database ────────────────────────────────────────────────────────────────────
+# Default: ./database/wv_property.db (relative to project root)
+# Railway: create a Volume → mount at /data → set DATABASE_PATH=/data/wv_property.db
+DATABASE_PATH=./database/wv_property.db
+
+# ── Admin panel ─────────────────────────────────────────────────────────────────
+ADMIN_PASSWORD=change-me-in-production
 SESSION_SECRET=change-me-in-production
 
-# ── Google OAuth2 (Gmail + Drive) ────────────────────────────
-# Create credentials at https://console.cloud.google.com/
-# 1. Enable Gmail API and Google Drive API
-# 2. Create an OAuth 2.0 Client ID (Desktop or Web application)
-# 3. Use the OAuth Playground (https://developers.google.com/oauthplayground)
-#    to generate a refresh token with these scopes:
-#      https://www.googleapis.com/auth/gmail.send
-#      https://www.googleapis.com/auth/drive.file
+# ── REST API auth ───────────────────────────────────────────────────────────────
+# Required to use POST/PUT/DELETE /api/properties and GET /api/contacts.
+# Leave blank in development to allow unauthenticated API access.
+API_KEY=your-secret-api-key
+
+# ── Gmail notifications (optional) ─────────────────────────────────────────────
+# Sends an email to NOTIFICATION_EMAIL every time a lead inquiry is submitted.
+# Setup: https://console.cloud.google.com → enable Gmail API → OAuth2 Client ID
+#   Use OAuth Playground to get a refresh token with scope:
+#   https://www.googleapis.com/auth/gmail.send
 GOOGLE_CLIENT_ID=
 GOOGLE_CLIENT_SECRET=
 GOOGLE_REFRESH_TOKEN=
+GOOGLE_GMAIL_USER=malickland@gmail.com
+NOTIFICATION_EMAIL=phil@malickland.net
 
-# The Gmail address that will send notification emails (must match the OAuth account)
-GOOGLE_GMAIL_USER=
-
-# Email address that will RECEIVE inquiry notifications
-NOTIFICATION_EMAIL=
-
-# Google Drive folder ID where property photo folders will be created
-# (Right-click a Drive folder → Get link → copy the ID from the URL)
+# ── Google Drive photo backup (optional) ────────────────────────────────────────
+# Uploads compressed photos to Drive after each upload.
+# Enable Google Drive API in the same Cloud project, add scope:
+#   https://www.googleapis.com/auth/drive.file
+# Get the folder ID from the Drive URL: drive.google.com/drive/folders/{ID}
 GOOGLE_DRIVE_FOLDER_ID=
+
+# ── AI listing content (optional) ───────────────────────────────────────────────
+# Powers /admin/ai/:id — generates a full marketing package per listing via GPT-4o.
+# Cost: ~$0.01–0.03 per generation.
+# Get key: https://platform.openai.com/api-keys
+OPENAI_API_KEY=
