+++ api/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node server.js",
     "dev": "nodemon server.js",
-    "test": "echo \"No tests configured\" && exit 0"
+    "test": "node --check server.js && echo 'Syntax OK'"
   },
   "keywords": [
     "real-estate",
@@ -17,19 +17,17 @@
   "license": "ISC",
   "type": "commonjs",
   "dependencies": {
-    "better-sqlite3": "^12.8.0",
-    "connect-redis": "^9.0.0",
+    "better-sqlite3": "12.8.0",
+    "better-sqlite3-session-store": "^0.1.0",
     "cors": "^2.8.6",
     "dotenv": "^17.4.1",
-    "express": "^5.2.1",
+    "express": "5.2.1",
     "express-rate-limit": "^8.3.1",
     "express-session": "^1.19.0",
-    "googleapis": "^171.4.0",
     "helmet": "^8.1.0",
     "morgan": "^1.10.1",
     "multer": "^2.1.1",
-    "redis": "^5.11.0",
-    "sharp": "^0.33.5"
+    "sharp": "0.33.5"
   },
   "devDependencies": {
     "nodemon": "^3.1.9"
