+++ app/listing.js
@@ -114,7 +114,7 @@ async function loadProperty() {
 }
 
 function escapeHtml(s) {
-  return String(s)
+  return String(s ?? '')
     .replace(/&/g, '&amp;')
     .replace(/</g, '&lt;')
     .replace(/>/g, '&gt;')
